Split file for later abstraction
