+++ R&M/R&M/CacheFramework/CachedThumbnail.swift
@@ -1,0 +1,12 @@
+//
+//  CachedThumbnail.swift
+//  R&M
+//
+//  Created by Mauricio Chirino on 03/01/21.
+//
+
+import Foundation
+
+public protocol CachedThumbnail {
+    var avatar: URL { get }
+}
