+++ R&M/R&M/CacheFramework/ThumbnailSource.swift
