+++ R&M/R&M/Infrastructure/DTOs/CharacterDTO.swift
@@ -7,18 +7,6 @@
 
 import Foundation
 
-public protocol CachedThumbnail {
-    var avatar: URL { get }
-}
-
-public protocol CardSourceable: CachedThumbnail {
-    var title: String { get }
-    var subtitle: String { get }
-    // TODO: these attributes below shouldn't be part of generic data
-    var species: String { get }
-    var origin: String { get }
-}
-
 struct CharacterLocationDTO: Codable {
     let name: String
 }
