+++ R&M/R&M/UIKit/UIConstants.swift
@@ -1,0 +1,12 @@
+//
+//  UIConstants.swift
+//  R&M
+//
+//  Created by Mauricio Chirino on 03/01/21.
+//
+
+import UIKit
+
+struct UIConstants {
+    static let standardPadding: CGFloat = 8
+}
